Removes `dtype=torch.long` declarations.
I know of no reason to think these are necessary.

Closes #236.

Closes #236.

Closes #236.

Closes #236.

+++ yoyodyne/models/hard_attention.py
@@ -81,9 +81,7 @@ class HardAttentionRNNModel(rnn.RNNModel):
         batch_size = encoder_mask.shape[0]
         decoder_hiddens = self.init_hiddens(batch_size)
         bos = (
-            torch.tensor(
-                [special.START_IDX], device=self.device, dtype=torch.long
-            )
+            torch.tensor([special.START_IDX], device=self.device)
             .repeat(batch_size)
             .unsqueeze(-1)
         )
